Reorganized the components
- Reorganized all of the components so we have a clean line between common (MainHeading, SubHeading and Paragraph) components and components representing actual chunks of content.
- Introduced consistency in the folder structure so that everything is using index.vue files and the references just use the paths.
- Switched all the references to be absolute from base, not long relative .. paths.
- Corrected filename on Chinese flag image
