+++ src/components/Funding/index.vue
@@ -15,9 +15,9 @@
 
 </template>
 <script>
-import MainHeading from '../Comman/Heading/mainHeading.vue';
-import SubHeading from '../Comman/Heading/subHeading.vue';
-import Paragraph from '../Comman/Paragraph';
+import MainHeading from '@/components/Comman/MainHeading';
+import SubHeading from '@/components/Comman/SubHeading';
+import Paragraph from '@/components/Comman/Paragraph';
 
 export default {
     name: "About_Funding",
